Repackage sui sdk
